Aggree that leaving a trailing / for root is OK
root_controller_spec.rb and sites_controller_spec.rb implemented two
conflicting requirements when deciding how to present the root of the
API depending on whether the root was fictionnal through API_MOUNT_PATH or the
real root of the service.

This commit chooses to allways present a trailing / when building hrefs

The related test in root_controller_spec has also been updated to
reflect that the tested behavior (prepending sid) is acheived by using
X-Api-Version header rather that X-Api-Path-Prefix on real deployments.