Merge implementations of status & status_for_disk
This makes it possible to only build the list of
active jobs with their associated resources once
instead of twice.

The updated status class method will take the types
of resources for which a status is required and build
an aggregated hash indexed by type for easy use by
controllers

The implementation is rather crude, using multiple
case statements instead of some form of generic
programming.