Use version of em-postgresl-adapter avoiding race
When the number of Fibers is greater than the number of connections to
postgresql in the connection Pool, a race condition occurs when
generating the SQL query as this requires a connection for encoding
context. Indeed the visitor used for Postgresql to generate the SQL from the AST
generated by active_record is not Fiber safe, and will not generate the
proper SQL if used concurrently by multiple Fibers. This will happen as
a connection must be available to escape Strings, and the generation
will hand over control to an other Fiber is none is available.