Fix latest commit retrieval for a resource
The Walker used to find the last commit for a path (resource) was
configured to sort the commit by date (commit time) and not
chronological order. As a more recent commit can have an inferior
date than a older commit, this sort method is not relevant and can
cause g5k-api to respond with incorrect (old) data.

This is weird that the bug was not encountered earlier, it's looks like
the sorting using Rugged::SORT_DATE is not deterministic. Indeed, a
specific commit triggered the bug in g5k-api, although it should have
happened earlier (two commits should have been sorted by date but were
not).

Fixes: bug #13541