Correctly handle 'now' and wrong advance reservations in job model
The job model was parsing the input job reservation date before passing
it to the internal OAR API. Parsing exceptions were caught to return nil,
so if the date was not parseable ('now' included which is a valid oarsub
keyword for advance reservations), resulting of job created as batch and
not advance reservations.

This is now fixed by not parsing the date, and by using the regex used
by oarsub to validate the advance reservation format (including 'now').
If the format is not valid, an error is returned.

Documentation was also updated to match oarsub behavior.