Move a part of Vlan API error management to kavlan lib
This align vlan api behavior with jobs api. Since g5k-api is a proxy in
those cases, we need to handle http errors from the backends (like
kavlan or oar). Internal libraries were created to make the call to the
backends, and it should be the responsability of those libraries to
raise errors when needed.